Rework DamageManagerBase into LifeStateComponent, which does all the calculations for health and armor

+++ Plugins/TrickyProperties/Source/TrickyProperties/Public/TrickyPropertiesLibrary.h
@@ -224,9 +224,9 @@ struct FTrickyPropertyInt
 		Value = MaxValue;
 	}
 
-	int32 GetNormalizedValue() const
+	float GetNormalizedValue() const
 	{
-		return MaxValue > 0 ? static_cast<float>(Value) / static_cast<float>(MaxValue) : 0;
+		return MaxValue > 0 ? static_cast<float>(Value) / static_cast<float>(MaxValue) : 0.f;
 	}
 
 	bool ReachedMinValue() const
